adding some events onValidationError and onValidationStart.
Also adding an array of elements on each observable so we can tell what elements if any that this observable is bound on. This is useful so we can hook into a validation error later and use an external lib like qtip to display the error messages instead of just inserting a span into the html.