Added clearAllErrors and validateAllObservables functions
This is for issue #431

clearAllErrors: Use in cases where you want to reset all of your
observable's validation errors to zero. Which is useful when you don't
want errors to show right away, until the user makes a change to the
property. This is particularly true when you have Async and Synchronous
rules on the same observable.

validateAllObservables: Use in cases when you want to force all the
validation to fire on your observables. Useful if you called
clearAllErrors, and you want to validate your properties before "save".